ipynb file containing all the analyses done
This file contains several blocks of code, together with charts and analyses, done for the technical challenge. Each block is slightly commented within the code, plus a global title and in some cases a small comment on the top of the block. At the beginning of the entire file, there are some general comments about the challenge.